More code review changes
