+++ Horizon/Horizon/Resources/Localizable.xcstrings
@@ -68,6 +68,9 @@
     "Learn" : {
 
     },
+    "Notebook" : {
+
+    },
     "Notebook for Course %@" : {
 
     },
