+++ Horizon/Horizon/Sources/Features/Notebook/Common/View/NotebookSearchBar.swift
@@ -16,14 +16,27 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
-struct NotebookCourse: Hashable {
-    let id: String
-    let course: String
-    let institution: String
+import Core
+import SwiftUI
 
-    init(from courseNote: CourseNote) {
-        id = courseNote.courseId
-        course = courseNote.course
-        institution = courseNote.institution
+struct NotebookSearchBar: View {
+
+    @Binding var term: String
+
+    var body: some View {
+        ZStack(alignment: .leading) {
+            TextField("",
+                  text: $term,
+                  prompt: Text(String(localized: "Search", bundle: .horizon))
+                )
+                .frame(height: 48)
+                .padding(.leading, 48)
+                .background(Color.white)
+                .clipShape(RoundedRectangle(cornerRadius: 32))
+                .shadow(color: Color.black.opacity(0.2), radius: 8, x: 1, y: 2)
+            Image(systemName: "magnifyingglass")
+                .foregroundColor(.textDarkest)
+                .padding(.leading, 16)
+        }
     }
 }
