+++ Horizon/Horizon/Sources/Features/Notebook/Course/NotebookCourseAssembly.swift
@@ -17,7 +17,6 @@
 //
 
 import Core
-import SwiftUICore
 
 final class NotebookCourseAssembly {
     static func make(courseId: String) -> CoreHostingController<NotebookCourseView>? {
