+++ Horizon/Horizon/Sources/Features/Notebook/Course/View/NotebookCourseViewModel.swift
@@ -17,7 +17,6 @@
 //
 
 import Core
-import SwiftUI
 import Combine
 
 @Observable
